examples/lot_sizing/requirements.txt
pandas
numpy
--pre
cosy-luigi
